requirements.txt
flask
psycopg2-binary
requests
google-generativeai
python-dotenv 